The user can now switch between editing mode and playing mode.
It initially starts in editing mode, where the user can switch the living state of a cell by clicking on it.

Once the user presses the spacebar, the program will go into playing mode, where the grid will update each cell according to the rules of Conway's Game of Life. The user can press spacebar again to return to editing mode.

In editing mode, the user has the option to press "E" to update the game by one step.